+++ DesktopClient/MainWindow.xaml.cs
@@ -33,7 +33,7 @@ namespace DesktopClient
         private void Button_Click(object sender, RoutedEventArgs e) {
             try {
                 string name = nameTextBox.Text;
-                double price = double.Parse(priceTextBox.Text);
+                decimal price = decimal.Parse(priceTextBox.Text);
                 int stock = Int32.Parse(stockTextBox.Text);
                 int minStock = Int32.Parse(minStockTextBox.Text);
                 int maxStock = Int32.Parse(maxStockTextBox.Text);
